+++ migrations/030-cleanup-unused-fields.js
@@ -1,0 +1,17 @@
+let db = require('../src/db').sequelize;
+
+module.exports = {
+  up: function() {
+    try {
+      return db.query(`
+        DROP TABLE meetings;
+        ALTER TABLE articles DROP duration;
+        ALTER TABLE articles DROP endDate;
+        ALTER TABLE ideas DROP duration;
+        ALTER TABLE ideas DROP endDate;
+			`);
+    } catch(e) {
+      return true;
+    }
+  }
+}