+++ src/lib/mail.js
@@ -18,7 +18,6 @@ var env = nunjucks.configure('email');
 var dateFilter   = require('../lib/nunjucks-date-filter');
 dateFilter.setDefaultFormat('DD-MM-YYYY HH:mm');
 env.addFilter('date', dateFilter);
-//env.addFilter('duration', duration);
 
 // Global variables.
 env.addGlobal('HOSTNAME', config.get('hostname'));