+++ src/models/Article.js
@@ -47,41 +47,6 @@ module.exports = function( db, sequelize, DataTypes ) {
 			}
 		},
 
-		endDate: {
-			type         : DataTypes.DATE,
-			allowNull    : true,
-			get          : function() {
-				var date = this.getDataValue('endDate');
-      },
-		},
-
-		endDateHumanized: {
-			type         : DataTypes.VIRTUAL,
-			get          : function() {
-				var date = this.getDataValue('endDate');
-				try {
-					if( !date )
-						return 'Onbekende datum';
-					return  moment(date).format('LLL');
-				} catch( error ) {
-					return (error.message || 'dateFilter error').toString()
-				}
-			}
-		},
-
-		duration: {
-			type         : DataTypes.VIRTUAL,
-			get          : function() {
-				if( this.getDataValue('status') != 'OPEN' ) {
-					return 0;
-				}
-
-				var now     = moment();
-				var endDate = this.getDataValue('endDate');
-				return Math.max(0, moment(endDate).diff(Date.now()));
-			}
-		},
-
 		sort: {
 			type         : DataTypes.INTEGER,
 			allowNull    : false,
@@ -260,11 +225,6 @@ module.exports = function( db, sequelize, DataTypes ) {
 		individualHooks: true,
 
 		validate: {
-			validDeadline: function() {
-				if( this.endDate - this.startDate < 43200000 ) {
-					throw Error('An article must run at least 1 day');
-				}
-			},
 			validExtraData: function(next) {
 
         let self = this;
@@ -482,7 +442,7 @@ module.exports = function( db, sequelize, DataTypes ) {
 						order = [['createdAt', 'DESC']];
 						break;
 					case 'date_asc':
-						order = [['endDate', 'ASC']];
+						order = [['startDate', 'ASC']];
 						break;
 					case 'date_desc':
 					default:
@@ -494,7 +454,7 @@ module.exports = function( db, sequelize, DataTypes ) {
 								WHEN 'DENIED'   THEN 0
 								                ELSE 1
 							END DESC,
-							endDate DESC
+							startDate DESC
 						`);
 
 				}
@@ -554,7 +514,7 @@ module.exports = function( db, sequelize, DataTypes ) {
 				order = [['createdAt', 'DESC']];
 				break;
 			case 'date_asc':
-				order = [['endDate', 'ASC']];
+				order = [['startDate', 'ASC']];
 				break;
 			case 'date_desc':
 			default:
@@ -566,7 +526,7 @@ module.exports = function( db, sequelize, DataTypes ) {
 								WHEN 'DENIED'   THEN 0
 								                ELSE 1
 							END DESC,
-							endDate DESC
+							startDate DESC
 						`);
 		}
 
@@ -748,13 +708,6 @@ module.exports = function( db, sequelize, DataTypes ) {
 						return site;
 					})
 					.then( site => {
-
-						// Automatically determine `endDate`
-						if( instance.changed('startDate') ) {
-							var duration = ( instance.config && instance.config.articles && instance.config.articles.duration ) || 90;
-							var endDate  = moment(instance.startDate).add(duration, 'days').toDate();
-							instance.setDataValue('endDate', endDate);
-						}
 
 						return resolve();